Use BottomSheetState; fix nullability & network
Replaced many rememberModalBottomSheetState usages with rememberBottomSheetState + SheetValue.Hidden across UI components. Consolidated broadcast receiver registration via androidx.core.content.ContextCompat.registerReceiver to remove SDK branching. Fixed nullability and I/O handling (removed unsafe ?. on okhttp response.body/byteStream/bytes where .use ensures non-null), adjusted song ID/selection types, lyrics parsing nullability, playlist import deserialization checks, and small math/type fixes. Added lint ignores and @SuppressLint for restricted APIs and fixed several localized string format issues.

+++ app/src/main/java/chromahub/rhythm/app/shared/presentation/components/common/ProgressBarStyles.kt
@@ -1754,8 +1754,8 @@ private fun DrawScope.createRoundedRectProgressPath(
             val dy = nextPoint.y - point.y
             val len = kotlin.math.sqrt(dx * dx + dy * dy).coerceAtLeast(0.001f)
             point = Offset(
-                point.x + (-dy / len) * wave.toFloat(),
-                point.y + (dx / len) * wave.toFloat()
+                point.x + (-dy / len) * wave,
+                point.y + (dx / len) * wave
             )
         }
